Date: change default calendar to Julian/Gregorian
Chose the Julian/Gregorian transition to match the UNIX default,
which is the September 1752 transition by Britain and its colonies.
This is implemented with a calendar compositor, that takes multiple
calendars and transition dates. So it's relatively easy to choose
a different transition date.

This code definitely needs some refactoring and cleanup. It could
also use more test cases.