[Hashes] Adds a comparator and changes the initializers a bit
Adds the IndifferentAccessComparator, that allows to also use symbols as
keys, even when the Hash was created with Strings as keys.

The default value in the actual initialize method was not being used at
all (probably a left over), so I remove it. I also added a few other
ways to initialize hashes, so that you most likely will be able to
initialize it any way you want without having to pass nils, just to
comply with the arity.