Thread.start raises SIGSEGV on Linux 64 bits
AFAIK `pthread_exit` raises an exception and forces an unwind, which
is catched by `__crystal_personality`. Trying to handle it, results
in a SIGSEGV because `ABI.unwind_get_language_specific_data(context)`
returns a pointer to 0x0.

I'm not sure this patch is adequate, or that I got the problem, but
all the specs still pass, and raising within a thread raises the
exception in the main thread when the thread is joined.