[Specs] Adds the concept of lexer objects
Lexer objects are based on the concept of page objects for integration
tests. They abstract the intricacies of testing the lexer, and provide a
DSL like syntax for the actual tests.

This serves the purposes of reducing boilerplate in tests, improving
tests readability, removing duplicate knowledge throughout the tests and
thus following SRP.

I extracted the lexer object into it's own class and file, because I
feel it will have different reasons to change than the actual tests.