String#camelcase handles -es
I hit this when I tried to init a new project with a name like
nifty-project. I ended up with:

```crystal
module Nifty-project

end
```

Not *positive* it makes sense to handle it in String#camelcase... but
really not sure there's a case where you would ever want a camelcased
string that looks like 'Nifty-project'. If so, maybe we can write a
String#constantize method or somethign similar.