pass flags when building the spec target
Previously, flags such as `--stats` would not be passed to the crystal
compiler when building spec with e.g. `make stats=1 spec`. This commit fixes
that. You can also easily build a version of `all_spec` in release mode, but
bc+obj codegen has taken 45 minutes (and is still going).