more correct regex for github repo validation
the old regex would recognize foo/bar/baz as valid repository as
slashes are arbitrary characters in the sense of the dot-symbol.
this new regex allows only one slash in the repo-string.
additionally i limited the regex to match line begin and ending,
otherwise foo/bar/baz would still pass as the regex matches the
foo/bar-portion of the string.