Yield start and finish indices in Readline.autocomplete
As of today, `Readline.autocomplete` only yields the text sent by
readline for autocompletion, which only contains the last token inside
the current buffer. For example:

  > abc[TAB]
  => will yield "abc"

  > abc [TAB]
  => will yield ""

  > abc def[TAB]
  => will yield "def"

Yielding the positional indices of the yielded string inside the current
line buffer is necessary so the library is usable for performing some
useful autocompletion.

Note that we could use `Readline.line_buffer` directly to guess the
position from where the autocompletion request originated, but 1/ it
wouldn't be 100% deterministic, and 2/ this buffer is not updated in
real-time, so it frequently contains the previous line read by readline
instead of the current line.