Do not delegate to LibSSL when there is nothing to read
In a HTTP::FixedLenghtContent that has an OpenSSL::SSL::Socket as it's
io, the connection was just hanging. This is because the last cycle of
the read loop would know that there are 0 remaining bytes to be read,
but the loop won't actually break until the io's read tells it it has
read 0 bytes.

Apparently, LibSSL does not like to be asked to read 0 bytes, which we
would do every time. So, the solution is just to not delegate to LibSSL
at all in that situation.