Use UInt32 to avoid overflow when Crypto::Bcrypt.cost = 31
When the cost is 31 (1 << 31 == 2147483648), signed Int32 overflows,
resulting in -2147483648. This results in zero key expansions when there
should be over 2 billion. Hashing the key this many times should take
days, if not weeks; instead, it takes less than a second. Users who
think they're setting the maximum cost, and therefore the highest
resilience to a brute-force attack, are actually getting the minimum,
even lower than MIN_COST.