Improve Range#sum performance
We can calculate the sum of `Range(Int)` by O(1) algorithm.
(This is the so-called finite arithmetic series.)

This implementation is a little dirty. However, this hack is indeed
only available on Int range, and this feature is hoped by many people
(See below URLs).

Warning: It has a tiny breaking change. If you want to calculate the sum
of a range that contains too small size Int to sum it up such like
`1..10_000_000.sum`, then its sum cause an integer overflow, and
probably its result is difference between the change.
But I think this change is no problem because such code is already buggy.

Reference URLs:

  - ActiveSupport's Range#sum:
    http://www.rubydoc.info/docs/rails/3.1.1/Range#sum-instance_method
  - Scale's Range#sum:
    http://www.scala-lang.org/news/2.10.4
    Above link says "Range#sum is now O(1)"