Fix to catch all exceptions in at_exit handlers.
`at_exit` handler is used for two purpose, I think: the first is to hook main
script, the another is to release global resource. Current
implementation does not catch all exceptions but the first exception
only and stop to execute handlers. It is wrong behavior in the second
usecase, because to release should execute even if raised an exception.