Use single colon for type declaration. Use `x = uninitialized T` to declare a variable without initializing it (unsafe). Fixes #758.
Using a single colon is easier to write and is consistent with the syntax for a type restriction in method arguments and return types. It also makes sense: `@x : Int32` effectively restricts the type of `@x` to be Int32. Then, using `x :: Int32` to mean "declare x to be an Int32 but don't initialize it's value" was confusing, because the meaning is totally different. That's why we changed it to be `x = uninitialized Int32`, which makes it super obvious that x is not initialized, and thus unsafe in some cases. The word "uninitialized" is long and hard to type, but its use is not frequent so it's OK.

The formatter is updated to automatically change `@x :: Int32` to `@x : Int32` or `@x = uninitialized Int32` depending on whether that is inside a method or at the class level, and changing `x :: Int32` to `x = uninitialized Int32`, so upgrading existing code is just about running `crystal tool format` on a project.