Renamed `inject` to `reduce`. Related to #1420.
The reason is that `inject` sounds like we are "injecting a binary operation" between all the elements, and in fact in Ruby you can do `e.inject(:+)`. However, this is not possible in Crystal.
Also, having `map` and `reduce` feels more natural than `map` and `inject`.
The idea of "reducing all elements into a single value" is maybe better.