String: handle printf's dynamic width and precision
Width field:
------------

The Width field specifies a minimum number of characters to output, and
is typically used to pad fixed-width fields in tabulated output, where
the fields would otherwise be smaller, although it does not cause
truncation of oversized fields.

The width field may be omitted, or a numeric integer value, or a dynamic
value when passed as another argument when indicated by an asterisk "`*`".
For example, `printf("%*d", 5, 10)` will result in "`   10`" being printed,
with a total width of 5 characters.

Though not part of the width field, a leading zero is interpreted as the
zero-padding flag mentioned above, and a negative value is treated as
the positive value in conjunction with the left-alignment "`-`" flag also
mentioned above.

Precision field:
----------------

The Precision field usually specifies a maximum limit on the output,
depending on the particular formatting type. For floating point numeric
types, it specifies the number of digits to the right of the decimal
point that the output should be rounded. For the string type, it limits
the number of characters that should be output, after which the string
is truncated.

The precision field may be omitted, or a numeric integer value, or a
dynamic value when passed as another argument when indicated by an
asterisk "`*`". For example, `printf("%.*s", 3, "abcdef")` will result in
"`abc`" being printed.

-- Source: https://en.wikipedia.org/wiki/Printf_format_string

Closes #2137