Use Deque for concurrency (scheduler, channels, readers, writers), and use `shift` instead of `pop`.
This makes this code:

~~~
10.times do |i|
  spawn do
    puts i
  end
end

Fiber.yield
~~~

print the numbers from 0 to 9 as one would expect. Right now they are printed in the reverse order.

It's also more fair for readers/writters waiting on a file descriptor because they will now be cycled.
For example this:

~~~
spawn { loop { puts 1 } }
spawn { loop { puts 2 } }
sleep 1
~~~

previously printed a short burst of 2, then only 1. Now it alternates between 1 and 2, which
feels more correct.