Promote results of +-*|&^ operations to 64-bit if any operand is 64-bit
In the case if both "a" and "b" are 64-bit variables, the expressions
like "(1 + a) * b" used to produce a 32-bit result. Which is the root
cause of rather annoying and hard to debug problems when doing 64-bit
integer calculations.

This patch promotes the result of the "+", "-", "*", "|", "&" and "^"
operations to a 64-bit type if one of the operands is 64-bit.

Fixes https://github.com/crystal-lang/crystal/issues/567