Compiler: guess type from Pointer.malloc and Pointer.null. Related to #2443
Pointer is a well-known built-in type with the difference that to create an
instance of it you can do Pointer(Int32).new(1234), and that creates a pointer
pointing at that address, but to allocate memory and get a pointer to it you
use Pointer.malloc. It makes sense that the language knows about this construct
(which is used everywhere) and guess the type from it. Pointer.null is another
useful variant so it's added as a rule too.

This removes some redundant type annotations in basic data structures (Array, Hash,
Deque, BitArray) and possibly other data structures in other shards, as well
as also removing type annotations where C bindings and pointers are involved.

This is the only exception needed, because other types always use `new` to
create themselves.