Compiler: don't implicitly infer instance var to be nilable if they are assigned in at least one initialize but not all of them.
This code is not clear:

```
class Foo
  def initialize(@x : Int32)
    @y = 1
  end

  def initialize
    @y = 2
  end
end
```

Is `@x` supposed to be nilable because it wasn't initialized in the second
'initialize'? Is it a mistake or an intention? It's not clear.

With this change, either an explicit type declaration like `@x : Int32?` or
assigning nil in the second initialize, `@x = nil` is required so this is
more explicit, there's no confusion, and if `@x` is used and wasn't supposed
to be nil in the first snippet the error is not obscure.