Replace Pygments with Rouge
Pygments in Jekyll is broken, in a rather cryptic way, for many
people. The recommendation is to switch to Rouge for syntax
highlighting:

https://github.com/jekyll/jekyll/issues/2604#issuecomment-49099487

There are claims by the authors of Jekyll that certain version
of Python will work, however many people in issue comments have
claimed this not to be the case.

https://github.com/jekyll/jekyll/search?utf8=%E2%9C%93&q=%22Failed+to+get+header%22&type=Issues

For this reason I believe the best course of action is to
change to Rogue, instead of instructing people to use old
versions of Python which may not fix the problem.