Compiler: revert tuple covariance.
This needs a lot more thought.
It's never correct to merge two different tuple types into a single one.
We might want to allow assigning a more specific tuple type to a less
specific one, but right now there's no code in the compiler to do such
thing, so this will have to be done in the future (if we ever decide to to it)