Compiler: stop tracking whether a type was allocated or not. Fixes #2607
Before 0.16.0 we needed to know whether a type was allocated or not because
if it wasn't we couldn't know the type of its instance variables and so all code
related to that type couldn't be codegen-ed. After 0.16.0 this is no longer true,
and code can always be generated, even if a type wasn't instantiated.