Improve error suggestion when using logical operators. Fixes #2715
When using `if foo and bar` the compiler was suggestion the use of
`rand` method, and that is because `rand` is defined at top level and
Crystal::MatchesLookup was finding it as the nearest option.

Now it checks the missing method is `and` or `or` and suggest the right
`&&` and `||` operators.