Added Digest::SHA256
The interface is the same as Digest::SHA1
A lot of code is duplicated between the two, maybe we should refactor
into a Digest::Utils module?
Or Digest::Utils::SHA for SHA-specific utils?
The padding function, for example, is the same between all SHAs.

Also, the Digest::* interface is probably gonna be implemented in the
same way for every hash function, only the digest method will change.
Maybe we should have an interface with hexdigest and base64digest
implemented (as a base class or mixin, I don't know if it's a thing in
crystal)?

Tests are also provided, based on the test suite for SHA1.