DNS: Addrinfo wrapper
Implements a wrapper over `LibC::Addrinfo` in order to reduce the
discrepancies between the `LibC::Sockaddr`, `LibC::SockaddrIn` and
`LibC::SockaddrIn6` structs for IPv4 and IPv6 addresses, and usage
of enums for the family, type and protocol values. The struct may
also be passed along and copied at will, which is complicated with
the different sockaddr C structs that are different sizes and suse
pointers to memory that is often freed.

The `Addrinfo` struct may be transparently used in place of
`LibC::Sockaddr`.