DNS: threaded resolver
The C `getaddrinfo` function is a blocking call, and thus will lock
the entire event loop whenever a Domain Name must be resolved. This
leads to terrible performance when a name takes many seconds to be
resolved or to timeout.

This patch spawns one (or many) threads the first time a Domain Name
must be resolved. These threads will then be responsible for calling
`getaddrinfo` and sending the response back to the pending Fiber,
avoiding to lock the event loop, yet still calling the original
syscall with all its extensions (eg: NSSwitch or mDNS).

The threadpool size defaults to 1, but can be adjusted by applications
before the threads are started, using `DNS.threadpool_size = 4` for
example or even disabled by setting it to `0`, in which case the
resolver will fallback to a blocking call.

The implementation is a little slow, and resolving "localhost" in a
tight loop results in half the performance from merely calling
`getaddrinfo` directly, but having many fibers resolving domains
concurrently should quickly overcome this performance hit. Please
note that memory usage is stable, and LLVM/BOEHM GC seem to correctly
free the memory allocated from the thread.