Harmonize UDP, TCP and UNIX sockets coding + documentation
Harmonize coding styles, variables naming, and methods called across
the different classes. Write missing documentation.

Add missing `open` (with arguments) and `accept?` methods to UNIXServer
so it's on par with TCPServer.

Disable ability to select a protocol for `UDPSocket.pair` since the
general methods (`new` and `open`) don't allow it, and OSes don't
support any other protocol than the default `IPPROTO_IP` (0).

Avoid to leak a file descriptor when creating a UNIXSocket or
UNIXServer when the path is too long than supported by the OS. Also
raise an ArgumentError exception instead of Exception.