Can't parse enum from upcase word
I'm not sure if this was the case before. I don't know if this is a bug or just misdocumented, but I expect there are not tests for it or else they would fail.

Due to the camelcasing, it won't convert `"BLUE"` into `"Blue"`, it outputs `"BLUE"` which is not found in the enum. However, using a lowercase `"blue"` works because that gives `"Blue"`.