Use addr2line tool to add file:line to backtraces (requires --debug)
This requires the executable to be built with `--debug` to include
the necessary information. Also requires the `addr2line` executable
to be installed on the machine running the binary, otherwise the
unhelpful `?:?` will be printed instead.

Somewhat demangles symbols to remove the pointers (`*` and `~`) as
well as the return type so it's easier to read.

Doesn't touch the `print_backtrace` methods that will continue to
report the raw backtrace.

closes #1270