Fix: libxml2 is likely to miss unwind tables on ARM
The `.ARM.extab` and `.ARM.exidx` unwind tables, specific to the ARM
architecture, aren't available unless we compile libxml2 with
`-fexception -funwind-tables`. Since they are likely to be missing,
and trying to raise from a context that is missing unwind tables
will fail, we instead memoize the exception, then try to raise it
from the program context, which will have the unwind tables.