Fix: AArch64 context switch (#3498)
Fiber context switch fails on AArch64 with a SIGBUS error. It's
caused by the stack being incorrectly aligned. Pushing any other
register to th stack fixes the issue.

It's also impossible to read a pointer value directly into the SP
register, so this patch avoids the pointer for the second argument
altogether. The same could be applied for all platforms.