Reset location after block yielding in macro expansion
Introduce #<loc:save> and #<loc:reset> new pragma comment, #<loc:save>
saves current location and #<loc:reset> resets it. They are useful to
keep original location when block yielding in macro expansion.

For example, this code makes broken error message:

    macro foo
      a = {{ yield }}
      a.ok
    end

    foo { 1 }

It cannot be compiled and error message is:

    in foo.cr:7: undefined method 'ok' for Int32

Error message points line 7 of foo.cr, but this code consists only 6 lines.

This bug comes from a location is not reset after block yielding in
macro expansion. For instance, `foo { 1 }` expands to:

    a =  begin #<loc:"foo.cr",6,7>1 end
    a.ok

So, the parser reports `a.ok` is on line 7 of foo.cr.

This commit fixes it by wrapping #<loc:save> and #<loc:reeset> to
yielding expanded code, such a like:

    a =  #<loc:save>begin #<loc:"foo.cr",6,7>1 end#<loc:reset>
    a.ok