Adds favicon to generated docs.
When I have a bunch of tabs open and I am trying to find docs pages it is hard
because there is no visible favicon.

This commit adds <link> elements to the <head> of the two docs HTML templates.
It uses the the main site favicon which, when generated and hosted at
crystal-lang.org, is located at the absolute path /images/favicon.{ico,png}.