Error when defining `finalize` on a struct. Fixes #3840
This is a breaking change and it also forced a rename of a valid method in LLVMDIBuilder.
However, we think it's better to choose another name in this case because without seeing
the struct's usage one could have confused that `finalize` method with a real finalizer,
so with a mistake.