Add `IO#peek`
This method lets you peek into possibly buffered data held by an IO.
With this, some methods, for example `gets`, can be performed more
efficiently: if we can see which bytes come next we can see if there's the
delimiter we are looking for.

Before this, this optimization for `gets` was implemented in `IO::Buffered`,
which has direct access to this buffer, with some other IO wrappers forwarding
their `gets` method to the wrapped IOs. By introducing `peek`, `gets` can now
be implemented directly in `IO`, by trying to peek into the buffer, if any.

Since IOs in Crystal are buffered in most cases (IO::FileDescriptor is, and IO::Memory
can return the whole bytes are in `peek`) it's almost always the case that
`gets` will benefit from this optimization. Additionally, `IO::Sized`, which
previously overrode `gets`, no longer needs to do it and can now just
override `peek`, which has a much simpler implementation. In this particular
case it also leads to a faster `gets` performance. This `peek` method can also be defined
for other IOs, like ARGF and HTTP::Content, simplifying their code.

Another future optimization is to enhance `IO::Delimtied` by trying to peek
into an IO's buffer to see if the delimiter is there, instead of working with
a byte buffer that has the size of the delimiter.

Finally, the current Zlib::Inflate implementation reads more data than it
should, which is incorrect and can lead to, for example, some zip files
to be incorrectly parsed. A correct solution to this is to either reimplement
the DEFLATE algorithm in Crystal, or to feed zlib data byte per byte. The
later is simpler, but slower. But with `peek` we can feed more data, know
how much was actually used, and only `skip` that much data without effectively
reading more than necessary.