Set: rename `merge!` to `concat`
Right now Set behaves like an Array without duplicates and it has almost the same set of methods as Array. It also has `merge!` where in Array it's called `concat`, so switching an object from Array to Set will cause compile errors. This makes the code easier to refactor. Besides, `merge` exists in Hash-like types, not in Array-like types.