Fix: configurable SO_REUSEPORT socket setting (disabled by default)
This led to confusing behavior when starting different TCP services
that default to the same default ports. They didn't fail to start,
but were failing to receive some requests.

The setting is now configurable and disabled by default. You may
enable it by setting the `reuse_port` named argument to true in the
following methods:

    TCPServer.new
    TCPServer.open
    HTTP::Server#bind
    HTTP::Server#listen