Push LibXML's global state to the list of GC roots
We are using the GC to allocate memory in LibXML, but it keeps a per-thread
global structure with pointers to dynamically allocated objects, which is
allocated using a regular malloc(). Push this structure to the GC's roots list
such that the GC knows about these pointers and doesn't free them on collection.