Skip macro method (#4237)
Introduces a skip macro method. When called, it stops macro expansion of the rest of the file. This is meant to be used to support Windows (and other target combinations) more cleanly: instead of interleaving code with different targets in the same file through {% if ... %}, diverging code can be kept in different files.