Fix formatting of String Array literals containing right parens
This fixes a bug where the formatter will automatically replace the
String Array literal tokens with parens regardless of whether the content
already contains a `)`.  In the case where a `)` already exists, the
resultant code is not compilable.  This patch keeps the author's original
choice of token if a `)` exists in the content, as they made a conscious
decision to choose that alternate token so that the code can compile.
Otherwise, the formatter will choose parens for code consistency as was
done previously.

For example,

```
%w{
  one)
  two)
}
```

would have previously had the curly braces replaced with parens, which
would no longer compile.  Now, it will stay with curly braces, and be
properly formatted.