Fully honor no color option when running specs
Spec runner have the option to disable colored output (ANSI codes) when
`--no-color` option is used, example:

    $ crystal spec -- --no-color

However, certain elements in the output didn't fully support that option,
resulting in a mix of non-color and color.

This change unifies that setting and ensures that `--no-color` usage
is fully honored by the output.

Ref #4292