Allow OAuth::Consumer to take TLS context
Addresses issue #4381

Oauth consumer objects determine their TLS configuration strictly
as a boolean, then pass it to HTTP clients which are capable of
more complex configuration for use cases such as ignoring peer
certificate validation, client certificate configuration, etc.

Add a tls parameter of Nil | Bool | OpenSSL::SSL::Context::Client,
checking for its presence to determine whether or not to fall
back to resolving the boolean value from the value of @scheme
when assigning the value for @tls, which remains a union of
Bool | OpenSSL::SSL::Context::Client which can be passed to the
HTTP client safely.