Rename HTTP.rfc1123_date to HTTP.format_date
The date format used for HTTP applications (such as cookies) according to [RFC 2612](https://tools.ietf.org/html/rfc2616#section-3.3.1)
is *not* an arbitrary [RFC 1123](https://tools.ietf.org/html/rfc1123#page-55) string but has a fixed timezone `GMT`. Therefore it
is not the same as `Time#to_rfc2822` and the method should not be called `rfc1123_date`.