Parser: prevent to finish a macro body when {% if/for/begin %} is nested (#4801)
Fixed #4769

This introduced `control_nest` counter variable to keep macro control
(if/for/begin) nesting information. When the parser find `end` inside
macro body, it checks `control_nest` and finishes a macro body if and
only if `control_nest` is `0` (and `nest` which is usual block nesting
information is `0` also).

And formatter is fixed also.