Use query string from URI-based instantiation
When instantiating a WebSocket using separate parameters for host and path, it is possible to pass in a query string to the connection by appending it to the end of the path parameter. This is effectively stripped out when is passed through as a URI since the `path` call on the URL does not return the query string.

Fixed by calling `full_path` instead of `path` on the given URI.