Use SipHash in Crystal::Hasher
Improves Crystal::Hasher to use the SipHash family of pseudo random
functions for improved security of hashes; for example protecting
against hash flooding DoS attacks.

Since the hashes are never disclosed to a potential attacker, we
rely on the faster 1-3 versions instead of the cryptographically
verified 2-4 versions.

Last but not least, this uses siphash-1-3 on 64-bit systems and
halfsiphash-1-3 on 32-bit systems, for best overall performance.