Add Time.monotonic to return monotonic clock
Relies on `mach_absolute_time` on macOS, which is more widely
available and is supposed to have a higher precision. Relies on
`clock_gettime(CLOCK_RELATIME)` on other POSIX platforms.

Exposes `Time.monotonic` as a `Time::Span` with unspecified
starting point, but ready for time computations and comparisons.

Also introduces a `Time.measure(&block)` sugar to measure how long a
block takes to run —using the monotonic clock, obviously.