Make String#[] not read out-of-bounds if string ends in unicode char (#5257)
* Char::Reader: Fail if multi-byte UTF-8 sequence is cut off prematurely

Before this change, Char::Reader#decode_char_at() wouldn't detect that
an UTF-8 sequence was cut off prematurely.  Instead, it continued on
reading, leading to an out-of-bounds read.

This fix should be regarded as band-aid for Char::Reader.

* String#[]: Detect out-of-bounds when ending in a UTF-8 character

As Char::Reader signals the end of a string by returning a NUL-char,
we need to check in String#[] if the read could go out of bounds.  This
only triggers if the first byte of the character would be OOB already,
which is fine.  If a later byte in the character sequence is missing,
Char::Reader would notice it.