Parser: fix the condition to decide var or call
There is source code `foo`, this `foo` is decided var or call on
parsing and the condition for this has a bug.

First, `foo(+1) is parsed as `foo + 1` if `foo` is variable.
Second, `case 1 when foo then 2` is parsed as `case 1 when foo() then 2`
even if `foo` is variable.

These bugs are fixed in this commit.