Don't print query params if they are empty
I found this while investigating this issue on Lucky:
https://github.com/luckyframework/lucky/issues/214

It turns out that because I parsed the query params in that handler it
set the URI#query to `""`. This caused `HTTP::Request#resource` to add
an extra `?` to the end of the path, even though the query params were
empty.

This PR fixes the issue by also checking for an empty string before
printing `?`.