Fix bcrypt hard limit on passwords to 71 bytes
Despite the original bcrypt paper claiming passwords must be a
maximum of 56 bytes, the implementations are compatible to up to 72
bytes.

Since increasing the limit doesn't break compatibility, but other
implementations allow as many as 72 bytes, let's increase the
arbitrary limitation of 51 characters (which was wrong anyway) to 72
bytes, minus the leading null byte, that is a password of 71 bytes.