Use shards version in crystal init tool
`shards version` is a new shards command which finds a `shard.yml` in the
directory tree above the path given and returns the version specified in it. We
can use it here to automatically generate the version constant for libraries
directly from shard.yml, meaning it will never be out of sync.

The `VERSION` constant is also moved into the main library file, instead of
being placed in a separate `version.cr`. The old `version.cr` convention
persisted from ruby due to the need for gemspec files to require only the
version constant from the library.