Improved search feature for API docs (#5229)
* Implement advanced search for API docs

This greatly improves client-side search on generated HTML pages of the API docs.
It uses the program description available as JSON data and implements a custom search algorithm.
The search results include types, constants, methods, macros matching the query string in
one or more of the fields name, namespace, description, arguments or type. A custom ranking algorithm
sorts results based on matching quality and relevance.

Search features:
* prefix filter: `#` matches instance methods, `.` matches class methods
and macros, `:` matches namespace
* search is generally case-insensitive but results with case-equality are
ranked higher.
* search results list can be navigated by key (when in focus):
  * `i`, `c`, `ArrowUp`: move upwards
  * `j`, `h`, `ArrowDown`: move downwards
  * `Enter`: load result
  * `Escape`: cancel search
  * arrows, `Enter` and `Escape` also work when the search-input is
  in focus
* support for URL fragment-based search query: `/#q=mysearch` will
initialize a search at page load
* page-wide shortkeys to start search: `s` or `/`

* Add jsonp support to allow search when API docs are loaded from local file system

* Add jsonp support to allow search when API docs are loaded from local file system

* Navigator: Fix keybindings, don't jump to result while searching

* Improve result ranking based on in-string match position

* Don't keep results list when clicking on result

When a search result is opened, the new page does not load the results list and the query field is not focused (`leaveSearchScope`)
To refine or start a new search, when the query field is focused (`q` or
`/` shortcut) it's content is selected.

* Fix: don't trigger global shortkeys when input has focus

* Add usage modal

* remove autofocus of search input

* Delay navigator event after search has been performed

* Keep current selected result centered on screen

* Remove scroll position persistance (interferes with results list)

* Delay execution of search until index is loaded

* Add custom implemention for centered scrolling position

`scrollIntoView({behaviour: smooth, block: center})` is only an experimental DOM feature and not supported by many browsers.