Add `Enum.valid?`
A valid enum value means that `ThisEnum.from_value?(value)` returns the
`ThisEnum` member, and an invalid enum value means that it returns
`nil`.
This is implemented as the such way.

The reason that this is implemented as a class method is explained
in NOTE section of the document comment.