Add boundary check for seconds in Time#initialize
Previously, `Time#add_span` did not handle times at the min or max range with positive or
negative offsets correctly because `@seconds` can legitimately be `< 0` or `> MAX_SECONDS`
when the offset is taken into account.
The boundary check was moved to the constructor to prevent manually
creating an invalid date.