[WIP] Fix: don't use O_NONBLOCK on STDIN, STDOUT and STDERR
Setting O_NONBLOCK doesn't only affect the file descriptor of the
application, but the whole file description which may be shared with
other programs. This is always the case for STDIN, STDOUT and STDERR
that can be changed at any time by other processes. This makes
Crystal applications a bad citizen too, since it changes those file
descriptors in unexpected ways.

This patch introduces a new class, IO::StdFileDescriptor. It still
relies on libevent to determine whether a file descriptor is ready,
but doesn't set O_NONBLOCK and instead uses blocking syscalls that
shall be interrupted with a timer triggering a SIGALRM signal, thus
achieving a (somewhat) nonblocking operation.

The SIGALRM signal must be trapped using sigaction to set the
SA_NODEFER and more importantly not set the SA_RESTART flag that
would prevent syscalls from failing with EINTR.

All syscalls that may be interrupted must now check both EAGAIN and
EINTR before deciding to retry or fail.