Avoid Nil for XML::Reader string getters
instead return an empty string if the methods are called in an invalid
reader state (before the first or after the last read).

A special case is the #value method, which could also return nil if
called on a node without a text value, like `<tag>`, but here an empty
string also makes sense.