Refactor URI::Parser removing obsolete step macro
This macro was previously used to spec internals of URI::Parser, but it
no longer works because the struct can't be inherited. It doesn't make
much sense to spec the internal API anyway. Having unit tests for
URI.parse is sufficient.
The specs for URI::Parser integrated into the specs for URI.parse.