Fix URI#to_s path starting with `/`
`#to_s` inserts a slash if `path` does not start with one and follows a
non-nil authority.
Paths starting with double slash (`"//"`) are prefixed by ("/.`) if it
does not follow an authority. Otherwise the URI would be considered
relative to the schema and the first path component interpreted as authority.