Fix: disable/restore signals when forking process
Always block signal reception before forking, even in fork/exec
cases, so the child will never receive a signal and wrongly forward
it to the parent process.

Always restores the sigmask after fork. During fork/exec the signal
pipe isn't replaced but signal handlers are resetted to `SIG_DFL` so
they won't be received anymore (and won't be forwarded to the
parent) before we restore the sigmask (inherited after exec).