Don't set std handles to nonblocking after fork
Even setting NONBLOCK for a split second is dangerous.
From what I can see the child's stdin/out/err is supposed to be
blocking at all times, so just reopen the original FD's in
blocking mode.

There is still a problem with handling closed filehandles that needs
to be solved. `crystal_program | fast_exiting_child` crashes.. but this
is apparently not a regression.