Semantic: correctly guess ivar type from splat argument (#6648)
* Semantic: correctly guess ivar type from splat argument

Before this commit the type of ivar `@foo` of `def initialize(*@foo : Int32)`
is guessed to `Int32`. It is very strange.

It should not be guessed to any types because this restriction means
variable-length tuple and we have no type for it. `Indexable(Int32)` is one
of generic types of variable-length tuple, but it is not perfect. So, it
should be specified explicitly if needed.

However, `@foo` type of `def initialize(*@foo : *{Int32})` can be guessed
to `{Int32}` easily. It should be guessed and it is useful for
`def initialize(*@foo : *T)` case.

A double-splat and ivar has the same problems. This commit fixes them.

* Use `same?` instead `==`

https://github.com/crystal-lang/crystal/pull/6648#discussion_r214559131