Revert "allow rand from zero value"
This reverts commit 6202e37b115904aee690b5cef63eddede165cc91.

That commit introduced a bad change: it silenced a valid error for an
arbitrary case.

`rand(n)` is always equivalent to `rand(0...n)`.

`rand(0...0)` even now says "Invalid range for rand: 0...0".

`rand(0)` asks for a number >= 0 and < 0. This makes no sense and should
obviously be an error for the same reason.