Treat untyped element in tuple and named tuple
Fixed #6718

The reason of #6718 is calling `node.update` with untyped `TupleLiteral` or
`NamedTupleLiteral`. An untyped node has no `node.program`, so it fails nil
assertion.

I think this `node.update` is not needed because `CleanupTransformer`
does not change the type of `node` execpt for untyped and unreachable
node.  However such nodes should be transformed to other node.
In other words `node.update` does nothing in general.

Thus, this replaces `node.update` with treating untyped element.