Configure Crystal compiler version in Makefile
Some complicated logic was used to detect the current compiler version
from git tags. This method is unreliable and could produce invalid
results.
This is completely removed from Crystal source code and the version is
set as env var `CRYSTAL_CONFIG_VERSION` in the `Makefile`. The value is
determined from the most recent release listed in `CHANGELOG.md`. This
is a reliable method even if the code is not checked out in a git
repository.
By default, `-dev` suffix is appended to the version. For creating a
release build with a non-dev version, `CRYSTAL_CONFIG_VERSION` needs
to be explicitly specified.