Introduce Fiber::Context struct
The fiber context holds the current stack top pointer (for a saved
stack), as well as a resumable flag that tells whether the fiber can
be resumed (its context was fully saved) or is currently running, or
is dead (its proc returned).

The resumable flag is required to prevent a fiber to be resumed
while it's still running. This should usually not happen with the
monothreaded scheduler, unless it's trying to resume a terminated
fiber. This will be very frequent with a multithreaded scheduler
where a thread B could try to resume a fiber that has just been
enqueued by thread A but didn't fully stored its context yet.

The resumable flag is also used to mark a fiber as dead, which means
that it can't be resumed anymore. In that case the scheduler should
raise an exception.