Thread-safe Channel (disabling 'select')
Channel now relies on Crystal::Mutex and Crystal::ConditionVariable.

'select' is disabled because mutex & condition variables assume that
a fiber can only be in a single waiting queue at a time (otherwise
it's suspended), but 'select' expects a single fiber to be queued
into many wait lists, which is impossible right now —and possibly
ever since it could lead to many threads concurrently trying to
resume the same fiber (oops).