MT: BDWGC safety
Uses a (multi)readers (single)writer lock to prevent BDWGC from
initiating a scan/collect while a scheduler thread is swapping
context, which could lead to report a mismatched stack pointer and
stack bottom for a thread, which would obviously segfault.

Implements a custom RWLock struct which is noticeably faster than
`pthread_rwlock` (on linux at least), and optimized to avoid
delaying a writer lock for too long, to avoid to much overhead to GC
collections.

Requires BDWGC to be patched with `bdwgc-set-stackbottom.patch` to
be capable to declare the stack bottom of any thread.