MT: safe channels (but disables select)
Channel now uses Crystal::Mutex and Crystal::ConditionVariable. It's
slow but safe.

`select` is disabled because mutex & condition variables assume that
a fiber can only be in a single waiting queue at a time (otherwise
it's suspended), but `select` expects a single fiber to be queued
into many wait lists, which is impossible.