MT: rework Crystal::EventLoop and IO::Evented
EventLoop now uses a single Event per Fiber, and declares a little
more generic API: `wait(io, :read | :write, timeout)` and
`sleep(time)`.

IO::Evented now initializes the Fiber Event for each Fiber that is
waiting for a file descriptor to be readable or writable (libevent2
allows multiple events to the same file descriptor), instead of
having a single event, which avoids an issue where concurrent fibers
could alter `#read_timeout` or `#write_timeout` and affect
previously set events.

IO::Evented is now thread-safe, and doesn't use a Deque anymore, but
Crystal::WaitDeque, a double-ended linked list of pending fibers —we
still need to keep a list of pending fibers to resume when an IO is
reopened or closed (EV_CLOSED is only available for the linux/epoll
backend).

The `sleep` methods now delegate to EventLoop, instead of Scheduler.

FIXME: Crystal::Event is kept as a struct to avoid a double
memory allocation (one for Crystal::Event and another one for
LibEvent2::Event) despite having ivars that can be modified from
different places which requires to have a direct access to the
struct through hackish `fiber.@event.as(Crystal::Event)` means. This
should be refactored, for example to have a single Crystal::Event
allocation with enough space to initialize a LibEvent2::Event field.