Use `Process.run` to avoid ugly output from `system`.
When the `system` command fails, it will print an ugly `--: 1: --: xdg-open: not found.`

This can be avoided by running the `Process.run` command without configuring it with an input, output or error. `system` configures it with an input, output or error as you can see here (it uses `Process::Redirect::Inherit`) but `Process.run` doesn't do that by default and discards the IO (it uses `Process::Redirect::Close`).

Co-Authored-By: joshuapinter <joshuapinter@gmail.com>