Benchmarks: switch, channel, mutex and http server
The switch, channel and mutex benchmarks are furious benchmarks that
will report and allow to compare the overhead of synchronization
primitives. They must be taken with a grain of salt, as they don't
represent any real world constraint.

Note that Fiber.yield with the current scheduler goes through the
event loop, which is not the case for the MT scheduler which
immediately switches fibers, so the switch benchmark is *much*
faster with MT and the actual difference with ST is the libevent
overhead!

The http server test is more interesting since it spawns fibers,
allocates memory (thus runs GC collections), involves the event loop
(thus libevent), and does some IO. It's still not representative of
a real world application, but gives a nice overview of the overhead
of MT, and the actual impact of changing or tweaking some internals.

NOTE: since Socket#accept is broken because of broken IO::Evented,
the http server benchmark doesn't currently run with MT and the
server will just hang.