Fix: protect against parallel resumes
A fiber may be enqueued twice and resumed from multiple threads in
parallel. A race condition is still present using resumable with
only two states, and lead to segfaults during a swapcontext.

The patch changes `Fiber::Context#resumable` to an atomic long, then
requires the scheduler to atomically change the resumable state to a
third value (2) to be capable to. The `Fiber#resumable?` and
`Fiber#running?` calls are now atomic to allow synchronizing threads
properly.