Introduce Fiber::Context struct
The fiber context holds the current stack top pointer (for a saved
stack), as well as a resumable flag that tells whether the fiber can
be resumed (its context was fully saved) or is currently running, or
is dead (its proc returned).

The resumable flag is required to prevent a fiber to be resumed
while it's still running. This should usually not happen with the
monothreaded scheduler, but still useful to find a buggy
synchronisation primitive that would enqueue the same fiber multiple
times (e.g. the current `select` implementation). This will be very
frequent with a multithreaded scheduler where a thread B could try
to resume a fiber that has just been enqueued by thread A but didn't
fully store its context yet.

The fiber status, running or dead, is monitored using an `@alive`
boolean. We can't set `context.resumable` to another value, because
the scheduler eventually swaps the context to another fiber, which
would overwrite `context.resumable`.

The scheduler now aborts with a fatal error whenever it would resume
a running or dead fiber, which would otherwise lead to a segfault.
It doesn't just ignore the fiber since to enqueue a fiber twice or a
dead fiber is an error that must be fixed.