Change String#to_i to parse octals with prefix 0o
This is a breaking change.
Previously, numbers starting with 0 would be parsed as octals.
A previous change in Crystal requires octals start with 0o.
Now that is not the case, and they will be parsed as base-10.
The 0o prefix must be present to treat it as an octal.
This addresses parsing 0 with prefix: true raising an error.

Additionally, YAML parsing used the old style.
This updates the spec to handle the new parsing.
There is no official documentation for YAML (that I'm aware of) for octals.